libalpm: don't chown download dir if not root
Commit 7bc5d55b56f41518e0a53eed13d4c523aea848e5 disbaled the chown after
downloading if the current user is not root. It only makes sense for us
to do the same on this side otherwise files would end up stuck as the
download user.

This is also important for other alpm tools that configure themselves
via pacman.conf but download to user directories.